Test using example projects
For now we test the :assemble task of the sample project with multiple Gradle
and Android Plugin versions.

The only way to communicate with the testing JVM is via system properties
(systemProperty); we use JSON in order to send structured data (the valid
combinations of Gradle and Android Plugin versions).

The example project grows a little more complex due to the need to accomodate
test requirements.